feat: Add post_id to blog post links in index.php
This commit modifies the index.php file to include the post_id in the blog post links. This change allows users to navigate to individual blog posts by clicking on the post titles. The post_id is retrieved from the database and appended to the URL in the href attribute of the anchor tag.

The purpose of this change is to improve the user experience by providing direct access to specific blog posts from the home page.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>