feat: Add comment functionality to post.php
This commit adds the ability for users to leave comments on individual blog posts. The code changes include adding a form for submitting comments, processing the form data, and inserting the comment into the database. After submitting a comment, the page is redirected back to the post with the comment section scrolled into view.

Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>