feat: Update blog post links in index.php with post_id
This commit modifies the index.php file to include the post_id in the blog post links. By appending the post_id to the URL in the href attribute of the anchor tag, users can now navigate directly to individual blog posts by clicking on the post titles.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>