feat: Add logout functionality to admin/functions.php
This commit adds a new function called logout to the admin/functions.php file. The function destroys the session and redirects the user to the appropriate page based on the current URL. If the current URL is 'admin', the user is redirected to '../index.php', otherwise, they are redirected to './index.php'.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>