feat: Update CSS styles for active links and add views column to view_all_posts.php
This commit adds a new CSS class called .active_link to the styles.css file. The class sets the background color of active links to black. This improves the visual indication of the currently active link.

Additionally, this commit modifies the view_all_posts.php file to include a new column for displaying the number of views for each post. This column provides valuable information about the popularity of each post.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>