feat: Add online users functionality to userMenu.php
This commit adds the necessary code to include online users functionality in the userMenu.php file. The code changes include checking for the 'onlineusers' query parameter and updating the active user count using AJAX. This feature allows users to see the number of active users on the website in real-time.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>