refactor: Remove commented out code for loading online users
The code changes in this commit involve removing the commented out code for loading online users in the scripts.js file. This code was previously used to fetch the number of online users using AJAX, but it is no longer needed. Removing this code improves code readability and maintainability.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring or cleanup.
Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>